+++ lib/model/payload/characteristic/characteristic.dart
@@ -1,9 +1,9 @@
-import 'package:health_kit_reporter/model/type/characteristic_type.dart';
-
 import '../../decorator/extensions.dart';
+import 'activity_move_mode.dart';
 import 'biological_sex.dart';
 import 'blood_type.dart';
 import 'fitzpatrick_skin_type.dart';
+import 'wheelchair_use.dart';
 
 /// Equivalent of [Characteristic]
 /// from [HealthKitReporter] https://cocoapods.org/pods/HealthKitReporter
@@ -20,19 +20,25 @@ class Characteristic {
     this.biologicalSex,
     this.birthday,
     this.bloodType,
-    this.skinType,
+    this.fitzpatrickSkinType,
+    this.wheelchairUseType,
+    this.activityMoveModeType,
   );
 
   final BiologicalSex? biologicalSex;
   final DateTime? birthday;
   final BloodType? bloodType;
-  final FitzpatrickSkinType? skinType;
+  final FitzpatrickSkinType? fitzpatrickSkinType;
+  final WheelchairUseType? wheelchairUseType;
+  final ActivityMoveModeType? activityMoveModeType;
 
   Map<String, String?> get map => {
         'biologicalSex': biologicalSex?.string,
         'birthday': birthday?.toIso8601String(),
         'bloodType': bloodType?.string,
-        'skinType': skinType?.string,
+        'fitzpatrickSkinType': fitzpatrickSkinType?.string,
+        'wheelchairUseType': wheelchairUseType?.string,
+        'activityMoveModeType': activityMoveModeType?.string,
       };
 
   /// General constructor from JSON payload
@@ -41,5 +47,10 @@ class Characteristic {
       : biologicalSex = BiologicalSexFactory.from(json['biologicalSex']),
         birthday = json['birthday']?.toString().date,
         bloodType = BloodTypeFactory.from(json['bloodType']),
-        skinType = FitzpatrickSkinTypeFactory.from(json['skinType']);
+        fitzpatrickSkinType =
+            FitzpatrickSkinTypeFactory.from(json['fitzpatrickSkinType']),
+        wheelchairUseType =
+            WheelchairUseTypeFactory.from(json['wheelchairUse']),
+        activityMoveModeType =
+            ActivityMoveModeTypeFactory.from(json['activityMoveMode']);
 }